Added random facts

+++ IsaiahJTurner/SKBounceAnimation.m
@@ -113,7 +113,7 @@ SKBounceAnimationStiffness SKBounceAnimationStiffnessHeavy  = .001f;
 }
 
 - (void) setNumberOfBounces:(NSUInteger)newNumberOfBounces {
-	[super setValue:[NSNumber numberWithUnsignedInt:newNumberOfBounces] forKey:@"numberOfBouncesKey"];
+	[super setValue:[NSNumber numberWithUnsignedInt:(int)newNumberOfBounces] forKey:@"numberOfBouncesKey"];
 	[self createValueArray];
 }
